+++ src/index.js
@@ -2,7 +2,7 @@ import PDFEdit from "./editor.js";
 import PdfViewerApi from "./api.js";
 
 exports.activate = function (context) {
-  PDFEdit.register();
+  PDFEdit.register(context);
   return {
     getV1Api: function () {
       return PdfViewerApi;